pkg/sensors: add reload helper function
This patch adds an updateSelectors function that can be used for the
reload functions of both generic tracepoints and kprobes.
